wip(client): 移动了onLoginInputClick
